+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-17 14:15:17
+
+### Changed Files
+- `src/features/admin/components/EditUserDialog.tsx`
+- `src/services/users/userService.ts`
+
+### Statistics
+```
+ 2 files changed, 76 insertions(+), 5 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-17 10:56:17
 
 ### Changed Files